feat: uncommit dispose to clear table after every test

+++ EFCoreRelationshipsPractice/Services/CompanyService.cs
@@ -27,7 +27,7 @@ namespace EFCoreRelationshipsPractice.Services
 
         }
 
-        public async Task<CompanyDto> GetById(long id)
+        public async Task<CompanyDto> GetById(int id)
         {
             throw new NotImplementedException();
         }
